+++ src/bitcoind.cpp
@@ -104,7 +104,7 @@ bool AppInit(int argc, char *argv[]) {
             return false;
         }
         try {
-            gArgs.ReadConfigFile(gArgs.GetArg("-conf", BITCOIN_CONF_FILENAME));
+            gArgs.ReadConfigFiles();
         } catch (const std::exception &e) {
             fprintf(stderr, "Error reading configuration file: %s\n", e.what());
             return false;